requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
python-decouple>=3.8
dj-database-url>=2.1
whitenoise>=6.6
cloudinary>=1.36
django-cloudinary-storage>=0.3
django-cors-headers>=4.3
Pillow>=10.0
